+++ r-package/tests/testthat/test_read_grid.R
@@ -1,14 +1,16 @@
 context("read_grid")
 
+# skip tests because they take too much time
+testthat::skip_on_cran()
+
 ### expected behavior ----------------
+  test_that("read_grid expected behavior", {
 
-test_that("read_grid expected behavior", {
+    # whole file
+    expect_true(is(  read_grid(city='nat', showProgress = FALSE), 'sf'))
+    expect_true(is(  read_grid(city='Natal', showProgress = FALSE), 'sf'))
 
-  # whole file
-  expect_true(is(  read_grid(city='nat', showProgress = FALSE), 'sf'))
-  expect_true(is(  read_grid(city='Natal', showProgress = FALSE), 'sf'))
-
-})
+  })
 
 
 ### expected errors and messages ----------------
@@ -24,8 +26,8 @@ test_that("read_grid errors and messages", {
   testthat::expect_error(read_grid())
 
   # Wrong geometry or showProgress
-  testthat::expect_error( read_grid(city = 'nat', showProgress = 'aaa'))
-  testthat::expect_error( read_grid(city = 'nat', showProgress = 333))
-  testthat::expect_error( read_grid(city = 'nat', showProgress = NULL))
+  testthat::expect_error( read_grid(city = 'rec', showProgress = 'aaa'))
+  testthat::expect_error( read_grid(city = 'rec', showProgress = 333))
+  testthat::expect_error( read_grid(city = 'rec', showProgress = NULL))
 
 })