+++ r-package/tests/testthat/test_read_landuse.R
@@ -1,5 +1,7 @@
 context("read_landuse")
 
+# skip tests because they take too much time
+testthat::skip_on_cran()
 
 ### expected behavior ----------------
   test_that("read_landuse expected behavior", {
@@ -19,10 +21,10 @@ test_that("read_landuse errors and messages", {
   testthat::expect_error(read_landuse(city = 'abc'))
 
   # Wrong year
-  testthat::expect_error(read_landuse(city = 'nat', year=1500))
+  testthat::expect_error(read_landuse(city = 'rec', year=1500))
 
   # Wrong geometry or showProgress
-  testthat::expect_error(read_landuse(city = 'nat', year=2019, geometry = 'aaa'))
-  testthat::expect_error(read_landuse(city = 'nat', year=2019, showProgress = 'aaa'))
+  testthat::expect_error(read_landuse(city = 'rec', year=2019, geometry = 'aaa'))
+  testthat::expect_error(read_landuse(city = 'rec', year=2019, showProgress = 'aaa'))
 
 })