+++ r-package/NEWS.md
@@ -1,12 +1,16 @@
 # log history of aopdata package development
 
+
+
 -------------------------------------------------------
 
-# aopdata v0.2.2 (dev)
+# aopdata v0.2.2
 
 **Minor changes**
 * Adds `crul` as a package dependency. Now aoptadata fails gracefully when there is internet connection problems. Closes #32.
-* `showProgress` now sets to `FALSE` in all tests and examples. Closes #32.
+* `showProgress` now sets to `FALSE` in all tests, examples and vignettes. Closes #32.
+* skip tests on CRAN. Tests are run only locally
+